+++ src/app/beer_garden/db/mongo/api.py
@@ -19,7 +19,6 @@ from mongoengine.queryset.visitor import Q, QCombination
 import beer_garden.db.mongo.models
 from beer_garden.db.mongo.models import MongoModel
 from beer_garden.db.mongo.parser import MongoParser
-from beer_garden.db.mongo.pruner import MongoPruner
 from beer_garden.db.mongo.util import (
     check_indexes,
     ensure_local_garden,
@@ -193,12 +192,12 @@ def initial_setup():
     ensure_users()
 
 
-def get_pruner():
-    return MongoPruner
+# def get_pruner():
+#     return MongoPruner
 
 
-def prune_tasks(**kwargs) -> Tuple[List[dict], int]:
-    return MongoPruner.determine_tasks(**kwargs)
+# def prune_tasks(**kwargs) -> Tuple[List[dict], int]:
+#     return MongoPruner.determine_tasks(**kwargs)
 
 
 def get_job_store():