+++ CHANGELOG.md
@@ -7,6 +7,7 @@ TBD
 - Added auto refresh checkbox to requests page
 - Added support for cross-server parent/child requests
 - Reuse existing rabbit admin queue by purging messages instead of creating new queue
+- Refactored Mongo pruner to use APScheduler
 - Fixed bug where Auto brew plugins did not parse NAME/VERSION properly from the beer.conf
 - Fixed bug where unresponsive Garden API was referencing a non existing configuration for default. New default is -1.
 - Fixed bug where Garden unresponsive metadata was cleared during child garden shutdown